+++ src/my_strlen.c
@@ -6,9 +6,9 @@
 
 #include "libmn.h"
 
-int	my_strlen(char *str)
+size_t	my_strlen(char *str)
 {
-	int i;
+	size_t i;
 	
 	i = 0;
 	while (str[i])
